bug fixed

+++ ME0StubFinder/src/PatUnitBeh.cc
@@ -34,7 +34,7 @@ int calculate_hit_count(const std::vector<uint64_t>& masked_data, bool light) {
 int calculate_layer_count(const std::vector<uint64_t>& masked_data) {
     int ly_count = 0;
     bool not_zero;
-    for (int d : masked_data) {
+    for (uint64_t d : masked_data) {
         not_zero =  (d!=0);
         ly_count += (int)not_zero;
     }
